Update README.md
I made significant edits and additions to this README to provide users a deeper understanding of what is in each data file and the the general parameters of the tests that led to each performance measurement. 

I also think we need to include a short table that simply provides the SCU to SCA mapping and a labeled diagram of the flight focal plane as a reference. I added a row to the "Included Files" to describe this file.

I reviewed the .ecsv files and the only issue I found is that the WFI_pixels_meet_requirements.ecsv appears to be missing  the two columns with the actual pixel counts and percentages described in the description?